third_party/WebKit/LayoutTests/http/tests/devtools/elements/move-node-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests elements drag and drop operation internals, verifies post-move selection.


Running: testDumpInitial
========= Original ========
- <div id="container">
      <div id="child1"></div>
      <div id="child2"></div>
      <div id="child3"></div>
      <div id="child4"></div>
  </div>

Running: testDragAndDrop
===== Moved child2 =====
- <div id="container">
      <div id="child1"></div>
      <div id="child3"></div>
      <div id="child2"></div>
      <div id="child4"></div>
  </div>
Selection: div#child2
